+++ projects/rocprofiler-compute/src/rocprof_compute_tui/analysis_tui.py
@@ -142,6 +142,7 @@ class tui_analysis(OmniAnalyze_Base):
             self.args.list_stats,
             self.args.filter_metrics,
             sys_info.iloc[0],
+            getattr(self, "_profiling_config", {}),
         )
         self.load_options(normalization_filter)